+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Adds `install_exclude_globs` to exclude certain files during installation of wheels.
+
 ### Changed
 
 - **BREAKING** Introduce `package_annotation` which replaces `always_build_packages`, `build_target_overrides`,